Update implementation of MockWindow#setImmediate
First of all, this never should have used the internal Node setImmediate
function, not sure why that was decided.

The new implementation will not behave exactly as setImmediate does in
Node and the browser. The difference is that with this implementation,
immediate callbacks are not guaranteed to run before timeouts and
interval callbacks.

The reason why we choosed this implementation is that otherwise the user
of this library would have to tick to make the immediate callback
run. We decided against this for simplicity.